remove flowconfigs resource handler for single active code paths
