+++ gobblin-restli/gobblin-flow-config-service/gobblin-flow-config-service-server/src/main/java/org/apache/gobblin/service/FlowExecutionResource.java
@@ -17,9 +17,19 @@
 
 package org.apache.gobblin.service;
 
+import java.time.Duration;
+import java.time.Instant;
+import java.util.Comparator;
+import java.util.Iterator;
 import java.util.List;
+import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.ObjectUtils;
+
+import com.google.common.base.Strings;
 import com.google.inject.Inject;
+import com.linkedin.data.template.SetMode;
+import com.linkedin.data.template.StringMap;
 import com.linkedin.restli.common.ComplexResourceKey;
 import com.linkedin.restli.common.EmptyRecord;
 import com.linkedin.restli.server.PagingContext;
@@ -35,15 +45,22 @@ import com.linkedin.restli.server.annotations.QueryParam;
 import com.linkedin.restli.server.annotations.RestLiCollection;
 import com.linkedin.restli.server.resources.ComplexKeyResourceTemplate;
 
+import lombok.extern.slf4j.Slf4j;
+
+import org.apache.gobblin.service.monitoring.FlowStatus;
+import org.apache.gobblin.service.monitoring.FlowStatusGenerator;
+import org.apache.gobblin.service.monitoring.JobStatusRetriever;
+
 
 /**
  * Resource for handling flow execution requests
  */
+@Slf4j
 @RestLiCollection(name = "flowexecutions", namespace = "org.apache.gobblin.service", keyName = "id")
 public class FlowExecutionResource extends ComplexKeyResourceTemplate<FlowStatusId, EmptyRecord, FlowExecution> {
 
   @Inject
-  FlowExecutionResourceHandler flowExecutionResourceHandler;
+  FlowExecutionResourceHandlerInterface flowExecutionResourceHandler;
 
   public FlowExecutionResource() {}
 
@@ -98,6 +115,158 @@ public class FlowExecutionResource extends ComplexKeyResourceTemplate<FlowStatus
   @Override
   public UpdateResponse delete(ComplexResourceKey<FlowStatusId, EmptyRecord> key) {
     return this.flowExecutionResourceHandler.delete(key);
+  }
+
+  public static List<FlowStatus> getLatestFlowStatusesFromGenerator(FlowId flowId,
+      Integer count, String tag, String executionStatus, FlowStatusGenerator flowStatusGenerator) {
+    if (count == null) {
+      count = 1;
+    }
+    log.info("get latest called with flowGroup " + flowId.getFlowGroup() + " flowName " + flowId.getFlowName() + " count " + count);
+
+    return flowStatusGenerator.getLatestFlowStatus(flowId.getFlowName(), flowId.getFlowGroup(), count, tag, executionStatus);
+  }
+
+  public static FlowStatus getFlowStatusFromGenerator(ComplexResourceKey<FlowStatusId, EmptyRecord> key,
+      FlowStatusGenerator flowStatusGenerator) {
+    String flowGroup = key.getKey().getFlowGroup();
+    String flowName = key.getKey().getFlowName();
+    long flowExecutionId = key.getKey().getFlowExecutionId();
+
+    log.info("Get called with flowGroup " + flowGroup + " flowName " + flowName + " flowExecutionId " + flowExecutionId);
+
+    return flowStatusGenerator.getFlowStatus(flowName, flowGroup, flowExecutionId, null);
+  }
+
+  /**
+   * Forms a {@link FlowExecution} from a {@link FlowStatus}
+   * @param monitoringFlowStatus
+   * @return a {@link FlowExecution} converted from a {@link FlowStatus}
+   */
+  public static FlowExecution convertFlowStatus(FlowStatus monitoringFlowStatus,
+      boolean includeIssues) {
+    if (monitoringFlowStatus == null) {
+      return null;
+    }
+
+    Iterator<org.apache.gobblin.service.monitoring.JobStatus> jobStatusIter = monitoringFlowStatus.getJobStatusIterator();
+    JobStatusArray jobStatusArray = new JobStatusArray();
+    FlowId flowId = new FlowId().setFlowName(monitoringFlowStatus.getFlowName())
+        .setFlowGroup(monitoringFlowStatus.getFlowGroup());
+
+    long flowEndTime = 0L;
+    long maxJobEndTime = Long.MIN_VALUE;
+    String flowMessage = "";
+
+    while (jobStatusIter.hasNext()) {
+      org.apache.gobblin.service.monitoring.JobStatus queriedJobStatus = jobStatusIter.next();
+
+      // Check if this is the flow status instead of a single job status
+      if (JobStatusRetriever.isFlowStatus(queriedJobStatus)) {
+        flowEndTime = queriedJobStatus.getEndTime();
+        if (queriedJobStatus.getMessage() != null) {
+          flowMessage = queriedJobStatus.getMessage();
+        }
+        continue;
+      }
+
+      maxJobEndTime = Math.max(maxJobEndTime, queriedJobStatus.getEndTime());
+
+      org.apache.gobblin.service.JobStatus jobStatus = new org.apache.gobblin.service.JobStatus();
+
+      Long timeLeft = estimateCopyTimeLeft(queriedJobStatus.getLastProgressEventTime(), queriedJobStatus.getStartTime(),
+          queriedJobStatus.getProgressPercentage());
+
+      jobStatus.setFlowId(flowId)
+          .setJobId(new JobId()
+              .setJobName(queriedJobStatus.getJobName())
+              .setJobGroup(queriedJobStatus.getJobGroup()))
+          .setJobTag(queriedJobStatus.getJobTag(), SetMode.IGNORE_NULL)
+          .setExecutionStatistics(new JobStatistics()
+              .setExecutionStartTime(queriedJobStatus.getStartTime())
+              .setExecutionEndTime(queriedJobStatus.getEndTime())
+              .setProcessedCount(queriedJobStatus.getProcessedCount())
+              .setJobProgress(queriedJobStatus.getProgressPercentage())
+              .setEstimatedSecondsToCompletion(timeLeft))
+          .setExecutionStatus(ExecutionStatus.valueOf(queriedJobStatus.getEventName()))
+          .setMessage(queriedJobStatus.getMessage())
+          .setJobState(new JobState()
+              .setLowWatermark(queriedJobStatus.getLowWatermark()).
+                  setHighWatermark(queriedJobStatus.getHighWatermark()));
+
+      if (includeIssues) {
+        jobStatus.setIssues(new IssueArray(queriedJobStatus.getIssues().get().stream()
+            .map(FlowExecutionResource::convertIssueToRestApiObject)
+            .collect(Collectors.toList())));
+      } else {
+        jobStatus.setIssues(new IssueArray());
+      }
+
+      if (!Strings.isNullOrEmpty(queriedJobStatus.getMetrics())) {
+        jobStatus.setMetrics(queriedJobStatus.getMetrics());
+      }
+
+      jobStatusArray.add(jobStatus);
+    }
+
+    flowEndTime = flowEndTime == 0L ? maxJobEndTime : flowEndTime;
+
+    jobStatusArray.sort(Comparator.comparing((org.apache.gobblin.service.JobStatus js) -> js.getExecutionStatistics().getExecutionStartTime()));
+
+    return new FlowExecution()
+        .setId(new FlowStatusId().setFlowGroup(flowId.getFlowGroup()).setFlowName(flowId.getFlowName())
+            .setFlowExecutionId(monitoringFlowStatus.getFlowExecutionId()))
+        .setExecutionStatistics(new FlowStatistics().setExecutionStartTime(getFlowStartTime(monitoringFlowStatus))
+            .setExecutionEndTime(flowEndTime))
+        .setMessage(flowMessage)
+        .setExecutionStatus(monitoringFlowStatus.getFlowExecutionStatus())
+        .setJobStatuses(jobStatusArray);
+  }
+
+  private static org.apache.gobblin.service.Issue convertIssueToRestApiObject(
+      org.apache.gobblin.runtime.troubleshooter.Issue issues) {
+    org.apache.gobblin.service.Issue converted = new org.apache.gobblin.service.Issue();
+
+    converted.setCode(issues.getCode())
+        .setSummary(ObjectUtils.firstNonNull(issues.getSummary(), ""))
+        .setDetails(ObjectUtils.firstNonNull(issues.getDetails(), ""))
+        .setSeverity(IssueSeverity.valueOf(issues.getSeverity().name()))
+        .setTime(issues.getTime().toInstant().toEpochMilli());
+
+    if (issues.getProperties() != null) {
+      converted.setProperties(new StringMap(issues.getProperties()));
+    } else {
+      converted.setProperties(new StringMap());
+    }
+
+    return converted;
+  }
+
+  /**
+   * Return the flow start time given a {@link FlowStatus}. Flow execution ID is
+   * assumed to be the flow start time.
+   */
+  private static long getFlowStartTime(FlowStatus flowStatus) {
+    return flowStatus.getFlowExecutionId();
+  }
+
+  /**
+   * Estimate the time left to complete the copy based on the following formula -
+   *  timeLeft = (100/completionPercentage - 1) * timeElapsed
+   * @param currentTime as an epoch
+   * @param startTime as an epoch
+   * @param completionPercentage of the job
+   * @return time left in seconds
+   */
+  public static long estimateCopyTimeLeft(Long currentTime, Long startTime, int completionPercentage) {
+    if (completionPercentage == 0) {
+      return 0;
+    }
+
+    Instant current = Instant.ofEpochMilli(currentTime);
+    Instant start = Instant.ofEpochMilli(startTime);
+    long timeElapsed = Duration.between(start, current).getSeconds();
+    return (long) (timeElapsed * (100.0 / (double) completionPercentage - 1));
   }
 }
