+++ gobblin-restli/gobblin-flow-config-service/gobblin-flow-config-service-server/src/main/java/org/apache/gobblin/service/FlowExecutionResourceHandlerInterface.java
@@ -25,16 +25,18 @@ import com.linkedin.restli.server.PagingContext;
 import com.linkedin.restli.server.UpdateResponse;
 
 
-public interface FlowExecutionResourceHandler {
+// Unlike FlowConfigsV2ResourceHandler, this is an interface rather than a class because it's implementation needs
+// classes from gobblin-service module, and adding gobblin-service as a dependency will cause circular dependency,
+public interface FlowExecutionResourceHandlerInterface {
   /**
    * Get {@link FlowExecution}
    */
-  public FlowExecution get(ComplexResourceKey<FlowStatusId, EmptyRecord> key);
+  FlowExecution get(ComplexResourceKey<FlowStatusId, EmptyRecord> key);
 
   /**
    * Get latest {@link FlowExecution}
    */
-  public List<FlowExecution> getLatestFlowExecution(PagingContext context, FlowId flowId, Integer count, String tag,
+  List<FlowExecution> getLatestFlowExecution(PagingContext context, FlowId flowId, Integer count, String tag,
       String executionStatus, Boolean includeIssues);
 
   /**
@@ -43,16 +45,16 @@ public interface FlowExecutionResourceHandler {
    * NOTE: `executionStatus` param not provided yet, without justifying use case, due to complexity of interaction with `countPerFlow`
    * and resulting efficiency concern of performing across many flows sharing the single named group.
    */
-  public List<FlowExecution> getLatestFlowGroupExecutions(PagingContext context, String flowGroup, Integer countPerFLow,
+  List<FlowExecution> getLatestFlowGroupExecutions(PagingContext context, String flowGroup, Integer countPerFLow,
       String tag, Boolean includeIssues);
 
   /**
    * Resume a failed {@link FlowExecution} from the point before failure
    */
-  public void resume(ComplexResourceKey<FlowStatusId, EmptyRecord> key);
+  void resume(ComplexResourceKey<FlowStatusId, EmptyRecord> key);
 
   /**
    * Kill a running {@link FlowExecution}
    */
-  public UpdateResponse delete(ComplexResourceKey<FlowStatusId, EmptyRecord> key);
+  UpdateResponse delete(ComplexResourceKey<FlowStatusId, EmptyRecord> key);
 }
