+++ gobblin-restli/gobblin-flow-config-service/gobblin-flow-config-service-server/src/test/java/org/apache/gobblin/service/FlowConfigResourceLocalHandlerTest.java
@@ -29,6 +29,7 @@ import com.linkedin.data.template.StringMap;
 
 import org.apache.gobblin.configuration.ConfigurationKeys;
 import org.apache.gobblin.runtime.api.FlowSpec;
+import org.apache.gobblin.service.modules.restli.FlowConfigsV2ResourceHandler;
 
 
 public class FlowConfigResourceLocalHandlerTest {
@@ -47,11 +48,11 @@ public class FlowConfigResourceLocalHandlerTest {
             setRunImmediately(true))
         .setProperties(new StringMap(flowProperties));
 
-    FlowSpec flowSpec = FlowConfigResourceLocalHandler.createFlowSpecForConfig(flowConfig);
+    FlowSpec flowSpec = FlowConfigsV2ResourceHandler.createFlowSpecForConfig(flowConfig);
     Assert.assertEquals(flowSpec.getConfig().getString(ConfigurationKeys.FLOW_GROUP_KEY), TEST_GROUP_NAME);
     Assert.assertEquals(flowSpec.getConfig().getString(ConfigurationKeys.FLOW_NAME_KEY), TEST_FLOW_NAME);
     Assert.assertEquals(flowSpec.getConfig().getString(ConfigurationKeys.JOB_SCHEDULE_KEY), TEST_SCHEDULE);
-    Assert.assertEquals(flowSpec.getConfig().getBoolean(ConfigurationKeys.FLOW_RUN_IMMEDIATELY), true);
+    Assert.assertTrue(flowSpec.getConfig().getBoolean(ConfigurationKeys.FLOW_RUN_IMMEDIATELY));
     Assert.assertEquals(flowSpec.getConfig().getString("param1"), "a:b:c*.d");
     Assert.assertEquals(flowSpec.getTemplateURIs().get().size(), 1);
     Assert.assertTrue(flowSpec.getTemplateURIs().get().contains(new URI(TEST_TEMPLATE_URI)));
@@ -64,12 +65,12 @@ public class FlowConfigResourceLocalHandlerTest {
         .setTemplateUris(TEST_TEMPLATE_URI).setSchedule(new Schedule().setCronSchedule(TEST_SCHEDULE).
             setRunImmediately(true))
         .setProperties(new StringMap(flowProperties));
-    flowSpec = FlowConfigResourceLocalHandler.createFlowSpecForConfig(flowConfig);
+    flowSpec = FlowConfigsV2ResourceHandler.createFlowSpecForConfig(flowConfig);
 
     Assert.assertEquals(flowSpec.getConfig().getString(ConfigurationKeys.FLOW_GROUP_KEY), TEST_GROUP_NAME);
     Assert.assertEquals(flowSpec.getConfig().getString(ConfigurationKeys.FLOW_NAME_KEY), TEST_FLOW_NAME);
     Assert.assertEquals(flowSpec.getConfig().getString(ConfigurationKeys.JOB_SCHEDULE_KEY), TEST_SCHEDULE);
-    Assert.assertEquals(flowSpec.getConfig().getBoolean(ConfigurationKeys.FLOW_RUN_IMMEDIATELY), true);
+    Assert.assertTrue(flowSpec.getConfig().getBoolean(ConfigurationKeys.FLOW_RUN_IMMEDIATELY));
     Assert.assertEquals(flowSpec.getConfig().getString("param1"),"value1");
     Assert.assertEquals(flowSpec.getConfig().getString("param2"),"value1-123");
     Assert.assertEquals(flowSpec.getConfig().getString("param3"), "a:b:c*.d");
