+++ gobblin-service/src/main/java/org/apache/gobblin/service/modules/core/GobblinServiceManager.java
@@ -64,11 +64,7 @@ import org.apache.gobblin.runtime.troubleshooter.MultiContextIssueRepository;
 import org.apache.gobblin.scheduler.SchedulerService;
 import org.apache.gobblin.service.FlowConfig;
 import org.apache.gobblin.service.FlowConfigV2Client;
-import org.apache.gobblin.service.FlowConfigsResource;
-import org.apache.gobblin.service.FlowConfigsResourceHandler;
 import org.apache.gobblin.service.FlowConfigsV2Resource;
-import org.apache.gobblin.service.FlowConfigsV2ResourceHandler;
-import org.apache.gobblin.service.FlowExecutionResourceHandler;
 import org.apache.gobblin.service.FlowId;
 import org.apache.gobblin.service.GroupOwnershipService;
 import org.apache.gobblin.service.Schedule;
@@ -77,6 +73,8 @@ import org.apache.gobblin.service.modules.db.ServiceDatabaseManager;
 import org.apache.gobblin.service.modules.orchestration.DagProcessingEngine;
 import org.apache.gobblin.service.modules.orchestration.Orchestrator;
 import org.apache.gobblin.service.modules.orchestration.UserQuotaManager;
+import org.apache.gobblin.service.modules.restli.FlowConfigsV2ResourceHandler;
+import org.apache.gobblin.service.modules.restli.FlowExecutionResourceHandler;
 import org.apache.gobblin.service.modules.scheduler.GobblinServiceJobScheduler;
 import org.apache.gobblin.service.modules.topology.TopologySpecFactory;
 import org.apache.gobblin.service.monitoring.DagManagementDagActionStoreChangeMonitor;
@@ -94,9 +92,6 @@ public class GobblinServiceManager implements ApplicationLauncher {
   // These two options are required to launch GobblinServiceManager.
   public static final String SERVICE_NAME_OPTION_NAME = "service_name";
   public static final String SERVICE_ID_OPTION_NAME = "service_id";
-
-  public static final String SERVICE_EVENT_BUS_NAME = "GobblinServiceManagerEventBus";
-
   private static final Logger LOGGER = LoggerFactory.getLogger(GobblinServiceManager.class);
   protected final ServiceBasedAppLauncher serviceLauncher;
   private volatile boolean stopInProgress = false;
@@ -119,11 +114,7 @@ public class GobblinServiceManager implements ApplicationLauncher {
 
   @Inject
   @Getter
-  protected FlowConfigsResourceHandler resourceHandler;
-
-  @Inject
-  @Getter
-  protected FlowConfigsV2ResourceHandler v2ResourceHandler;
+  protected FlowConfigsV2ResourceHandler resourceHandler;
 
   @Inject
   @Getter
@@ -289,7 +280,7 @@ public class GobblinServiceManager implements ApplicationLauncher {
   private void configureServices(){
     if (configuration.isRestLIServerEnabled()) {
       this.restliServer = EmbeddedRestliServer.builder()
-          .resources(Lists.newArrayList(FlowConfigsResource.class, FlowConfigsV2Resource.class))
+          .resources(Lists.newArrayList(FlowConfigsV2Resource.class, FlowConfigsV2Resource.class))
           .injector(injector)
           .build();
 
@@ -318,10 +309,8 @@ public class GobblinServiceManager implements ApplicationLauncher {
       this.gitConfigMonitor.setActive(true);
     }
 
-    // Announce to d2 after services are initialized regardless of leadership if configuration is not enabled
-    if (!this.configuration.isOnlyAnnounceLeader()) {
-      this.d2Announcer.markUpServer();
-    }
+    // Announce to d2 after services are initialized
+    this.d2Announcer.markUpServer();
 
     // Populate TopologyCatalog with all Topologies generated by TopologySpecFactory
     // This has to be done after the topologyCatalog service is launched
@@ -359,9 +348,7 @@ public class GobblinServiceManager implements ApplicationLauncher {
     this.stopInProgress = true;
     try {
       // Stop announcing GaaS instances to d2 when services are stopped
-      if (!configuration.isOnlyAnnounceLeader()) {
-        this.d2Announcer.markDownServer();
-      }
+      this.d2Announcer.markDownServer();
       this.serviceLauncher.stop();
     } catch (ApplicationException ae) {
       LOGGER.error("Error while stopping Gobblin Service Manager", ae);
