+++ gobblin-service/src/test/java/org/apache/gobblin/service/GobblinServiceManagerTest.java
@@ -70,6 +70,7 @@ import org.apache.gobblin.service.modules.orchestration.MysqlDagActionStoreTest;
 import org.apache.gobblin.service.modules.orchestration.MysqlDagStateStore;
 import org.apache.gobblin.service.modules.orchestration.MysqlMultiActiveLeaseArbiterTest;
 import org.apache.gobblin.service.modules.orchestration.ServiceAzkabanConfigKeys;
+import org.apache.gobblin.service.modules.restli.FlowConfigsV2ResourceHandler;
 import org.apache.gobblin.service.modules.utils.FlowCompilationValidationHelper;
 import org.apache.gobblin.service.monitoring.DagManagementDagActionStoreChangeMonitor;
 import org.apache.gobblin.service.monitoring.FsJobStatusRetriever;
@@ -336,11 +337,11 @@ public class GobblinServiceManagerTest {
     FlowConfig uncompilableFlowConfig = new FlowConfig().setId(UNCOMPILABLE_FLOW_ID).setTemplateUris(TEST_TEMPLATE_URI)
         .setSchedule(new Schedule().setCronSchedule(TEST_SCHEDULE).setRunImmediately(true))
         .setProperties(new StringMap(flowProperties));
-    FlowSpec uncompilableSpec = FlowConfigResourceLocalHandler.createFlowSpecForConfig(uncompilableFlowConfig);
+    FlowSpec uncompilableSpec = FlowConfigsV2ResourceHandler.createFlowSpecForConfig(uncompilableFlowConfig);
     FlowId flowId = createFlowIdWithUniqueName(TEST_GROUP_NAME);
     FlowConfig runOnceFlowConfig = new FlowConfig().setId(flowId)
         .setTemplateUris(TEST_TEMPLATE_URI).setProperties(new StringMap(flowProperties));
-    FlowSpec runOnceSpec = FlowConfigResourceLocalHandler.createFlowSpecForConfig(runOnceFlowConfig);
+    FlowSpec runOnceSpec = FlowConfigsV2ResourceHandler.createFlowSpecForConfig(runOnceFlowConfig);
 
     // add the non compilable flow directly to the spec store skipping flow catalog which would not allow this
     this.gobblinServiceManager.getFlowCatalog().getSpecStore().addSpec(uncompilableSpec);
@@ -377,7 +378,7 @@ public class GobblinServiceManagerTest {
     Assert.assertEquals(specs.size(), 0);
   }
 
-  @Test //(dependsOnMethods = "testRestart")
+  @Test (dependsOnMethods = "testRestart")
   public void testUncompilableJob() throws Exception {
     FlowId flowId = new FlowId().setFlowGroup(TEST_GROUP_NAME).setFlowName(MockedSpecCompiler.UNCOMPILABLE_FLOW);
     URI uri = FlowSpec.Utils.createFlowSpecUri(flowId);
